토큰생성방법 groupid, userid로 변경

+++ src/main/java/com/intern_project/user/mapper/UserMapper.java
@@ -19,7 +19,7 @@ public interface UserMapper {
     //이메일로 계정찾기
     UserGroup findUserGroupByEmail(@Param("email") String email);
     //groupid로 계정찾기
-    UserGroup findUserGroupById(@Param("id") Long id);
+    UserGroup findUserGroupById(@Param("id") int id);
